+++ liyingtao/point.h
@@ -15,18 +15,18 @@ private:
 	long double z = 0;
 
 public:
-	double getX() const;
-	void setX(double x);
-	double getY() const;
-	void setY(double y);
-	double getZ() const;
-	void setZ(double z);
+	long double getX() const;
+	void setX(long double x);
+	long double getY() const;
+	void setY(long double y);
+	long double getZ() const;
+	void setZ(long double z);
 
 public:
 	Point() = default;
-	Point(double x, double y, double z);
+	Point(long double x, long double y, long double z);
 	void setPoint(long double x, long double y, long double z);
-	void movePoint(double x, double y, double z);
+	void movePoint(long double x, long double y, long double z);
 	void movePoint(const Vector &vector);
 };
